software/hive/backend/pyproject.toml
[project]
name = "hive-backend"
version = "0.1.0"
description = "Hive backend API"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy>=2.0",
    "alembic>=1.14",
    "psycopg2-binary>=2.9",
    "pyjwt>=2.9",
    "bcrypt>=4.2",
    "python-multipart>=0.0.18",
    "pydantic-settings>=2.7",
    "pydantic[email]>=2.7",
    "slowapi>=0.1.9",
    "cryptography>=45.0",
    "requests>=2.32",
]
